requirements.txt
streamlit
tensorflow
gdown
Pillow
numpy

